moved Swift Package test to integration-tests
