+++ Cargo.toml
@@ -34,7 +34,9 @@ members = [
   "crates/swift-bridge-ir",
   "crates/swift-bridge-macro",
 
-  "crates/swift-integration-tests","crates/swift-bridge-build/tests/test-swift-packages",
+  "crates/swift-integration-tests",
+  "SwiftRustIntegrationTestRunner/integration-test-create-swift-package",
+  "SwiftRustIntegrationTestRunner/swift-package-rust-library-fixture",
 
   "examples/async-functions",
   "examples/codegen-visualizer",
