+++ SwiftRustIntegrationTestRunner/integration-test-create-swift-package/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "integration-test-create-swift-package"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+swift-bridge-build = { path = "../../crates/swift-bridge-build" }
