+++ SwiftRustIntegrationTestRunner/swift-package-rust-library-fixture/build.sh
