+++ SwiftRustIntegrationTestRunner/swift-package-test-package/.gitignore
@@ -5,5 +5,3 @@
 xcuserdata/
 DerivedData/
 .swiftpm/xcode/package.xcworkspace/contents.xcworkspacedata
-.vscode
-.swiftpm
