--- crates/swift-bridge-build/tests/test-project/Sources/test_project/main.swift
@@ -1,3 +1,0 @@
-import MySwiftPackage
-
-print(hello_rust().toString())
