+++ test-integration.sh
@@ -13,6 +13,27 @@ cd $ROOT_DIR
 
 cd SwiftRustIntegrationTestRunner
 
+# Delete previous generated files/folders
+rm -r swift-package-rust-library-fixture/generated || true
+rm -r swift-package-rust-library-fixture/MySwiftPackage || true
+rm -r swift-package-rust-library-fixture/target || true
+rm -r swift-package-test-package/.build || true
+
+# Build Rust
+mkdir swift-package-rust-library-fixture/generated
+
+./swift-package-rust-library-fixture/build.sh
+
+# Create Swift Package
+cargo run -p integration-test-create-swift-package
+
+# Test Swift Package
+echo "TESTING SWIFT PACKAGE"
+cd swift-package-test-package
+swift test
+cd ..
+
+
 xcodebuild \
   -project SwiftRustIntegrationTestRunner.xcodeproj \
   -scheme SwiftRustIntegrationTestRunner \
